HevcSei: stop parsing on payload size exceeding remaining bits
HevcSei.cpp compares m_payloadSize, which is a byte count, against
AvailableBits(), a bit count, so a payload up to 8x the remaining data
passes the check. SkipBits() then runs past the end and the parse loop
can keep going on garbage instead of stopping. Compare against
AvailableBits() / 8 instead.

AvailableBits() itself underflows to a huge unsigned value once
m_posBits passes length * 8; clamp it to 0.

m_payloadType is accumulated in 255 increments, so widen it from
uint8_t to uint32_t to match m_payloadSize.

Needs a malformed stream to trigger. Not exercised on hardware.

+++ xbmc/utils/HevcSei.h
@@ -30,7 +30,7 @@ public:
   CHevcSei() = default;
   ~CHevcSei() = default;
 
-  uint8_t m_payloadType{0};
+  uint32_t m_payloadType{0};
   size_t m_payloadSize{0};
 
   // In relation to the input SEI rbsp payload
